Fixing a small error in resultsGUI.py which cause crash when a non-indexed area in interaction energy bar plot is clicked on
Former-commit-id: 0d7fa5b1acfebd8779db50ae69d28f61182f1c16

+++ source/calcGUI.py
@@ -221,10 +221,11 @@ class DesignInteractCalculate(QtWidgets.QMainWindow,calcGUI_design.Ui_MainWindow
 				if buttonReply == QtWidgets.QMessageBox.No:
 					pass
 				elif buttonReply == QtWidgets.QMessageBox.Yes:
-					rmtree(self.calcParams.outFolder,ignore_errors=True)
+					if os.path.exists(self.calcParams.outFolder):
+						rmtree(self.calcParams.outFolder,ignore_errors=True)
 			elif self.calcParams.calcState in ['Filtering','Calc']:
-				#pass
-				rmtree(self.calcParams.outFolder,ignore_errors=True)
+				if os.path.exists(self.calcParams.outFolder):
+					rmtree(self.calcParams.outFolder,ignore_errors=True)
 
 			# Return True in case this was called by exit window.
 			return True
